refactor(onboarding):  create page widgets
