+++ lib/ui/pages/onboarding/widgets/onboarding_slide.dart
@@ -1,0 +1,50 @@
+import 'package:flutter/material.dart';
+import 'package:gap/gap.dart';
+
+class OnboardingSlide extends StatelessWidget {
+  const OnboardingSlide({
+    required this.image,
+    required this.title,
+    required this.description,
+    super.key,
+  });
+
+  final String image;
+  final String title;
+  final String description;
+
+  @override
+  Widget build(BuildContext context) {
+    final size = MediaQuery.sizeOf(context);
+
+    return Column(
+      children: [
+        Gap(36),
+        SizedBox(
+          height: size.height * 0.4,
+          child: Image.asset(image),
+        ),
+        Gap(48),
+        Text(
+          title,
+          style: Theme.of(context).textTheme.headlineLarge!.copyWith(
+                color: Colors.white,
+                fontWeight: FontWeight.bold,
+              ),
+        ),
+        Gap(24),
+        Padding(
+          padding: const EdgeInsets.symmetric(horizontal: 24),
+          child: Text(
+            description,
+            textAlign: TextAlign.center,
+            style: Theme.of(context).textTheme.bodyLarge!.copyWith(
+                  color: Colors.white,
+                  fontWeight: FontWeight.w500,
+                ),
+          ),
+        ),
+      ],
+    );
+  }
+}
